part price is now double

+++ NolekAPI/Model/tblParts.cs
@@ -7,6 +7,6 @@ namespace NolekAPI.Model
         public int PartID { get; set; }
         public string PartName { get; set; }
         public int NumberInStock { get; set; }
-        public int PartPrice { get; set; }
+        public double PartPrice { get; set; }
     }
 }
